+++ features/mpaccount/src/main/java/com/jpp/mpaccount/account/lists/UserMovieListType.kt
@@ -1,0 +1,14 @@
+package com.jpp.mpaccount.account.lists
+
+import androidx.annotation.StringRes
+import com.jpp.mpaccount.R
+
+/**
+ * Represents the type of user movie list that the application can
+ * show.
+ */
+internal enum class UserMovieListType(@StringRes val titleRes: Int) {
+    FAVORITE_LIST(R.string.user_account_favorite_title),
+    RATED_LIST(R.string.user_account_rated_title),
+    WATCH_LIST(R.string.user_account_watchlist_title)
+}
